Made code more readable
